# spec/api.yaml
...
openapi: 3.0.3
info:
  title: TUD Assistance Backbone API
  description: Component for analyzing learning process data in the form of xAPI statements and providing feedback to
    learners' interactions and assistance and suggestions with regard to the user's learning state and the corresponding
    learning content.
  license:
    name: GNU General Public License
    url: https://www.gnu.org/licenses/old-licenses/gpl-3.0.de.html
  version: 1.0.0
...
security:
- bearerAuth: []
paths:
...
  /api/v1/lco:
    get:
      tags:
      - Expert Module
      summary: Get Learning Content Objects
      description: Get all Learning Content Objects.
      operationId: getLearningContentObjects
      parameters:
      - $ref: '#/components/parameters/pageNumber'
      - $ref: '#/components/parameters/numberOfObjectsPerPage'
      responses:
        200:
          description: Successfully retrieved Learning Content Objects
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LearningContentObjectList'
        400:
          description: Bad request
        401:
          description: Authentication failed
    post:
      tags:
      - Expert Module
      summary: Create Learning Content Object
      description: Create a Learning Content Object.
      operationId: createLearningContentObject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreliminaryLearningContentObject'
      responses:
        201:
          description: Successfully created Learning Content Object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LearningContentObject'
        400:
          description: Bad request
        401:
          description: Authentication failed
        409:
          description: LCO with object ID already exists
        422:
          description: Unprocessable entity
      x-codegen-request-body-name: body
  /api/v1/lco/search:
    post:
      tags:
      - Expert Module
      summary: Search for Learning Content Objects
      description: Search for Learning Content Objects by attributes.
      operationId: searchForLearningContentObjects
      parameters:
      - $ref: '#/components/parameters/pageNumber'
      - $ref: '#/components/parameters/numberOfObjectsPerPage'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/LearningContentObjectAttributeSearchParameter'
      responses:
        200:
          description: Successfully retrieved Learning Content Objects
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LearningContentObjectList'
        400:
          description: Bad request
        401:
          description: Authentication failed
        422:
          description: Unprocessable entity
      x-codegen-request-body-name: body
  /api/v1/lco/{lcoId}:
    get:
      tags:
      - Expert Module
      summary: Get Learning Content Object
      description: Get a Learning Content Object with the specified ID.
      operationId: getLearningContentObject
      parameters:
      - $ref: '#/components/parameters/lcoId'
      responses:
        200:
          description: Successfully retrieved Learning Content Object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LearningContentObject'
        401:
          description: Authentication failed
        404:
          description: A Learning Content Object with the specified ID does not exist
    put:
      tags:
      - Expert Module
      summary: Update Learning Content Object
      description: Update a Learning Content Object with the specified ID.
      operationId: updateLearningContentObject
      parameters:
      - $ref: '#/components/parameters/lcoId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreliminaryLearningContentObject'
      responses:
        200:
          description: Successfully updated Learning Content Object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LearningContentObject'
        400:
          description: Bad request
        401:
          description: Authentication failed
        404:
          description: A Learning Content Object with the specified ID does not exist
        422:
          description: Unprocessable entity
      x-codegen-request-body-name: body
    patch:
      tags:
      - Expert Module
      summary: Patch Learning Content Object
      description: Update certain parts of a Learning Content Object with the specified ID. The existing list of Learning Content Object attributes updated by applying the provided patches. If the list of attributes should be overwritten, the update operation has to be used. The patches are applied in the provided order.
      operationId: patchLearningContentObject
      parameters:
      - $ref: '#/components/parameters/lcoId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/LearningContentObjectPatch'
      responses:
        200:
          description: Successfully patched Learning Content Object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LearningContentObject'
        400:
          description: Bad request
        401:
          description: Authentication failed
        404:
          description: A Learning Content Object with the specified ID does not exist
        422:
          description: Unprocessable entity
      x-codegen-request-body-name: body
    delete:
      tags:
      - Expert Module
      summary: Delete Learning Content Object
      description: Delete a Learning Content Object with the specified ID.
      operationId: deleteLearningContentObject
      parameters:
      - $ref: '#/components/parameters/lcoId'
      responses:
        204:
          description: Successfully deleted Learning Content Object
        401:
          description: Authentication failed
        404:
          description: A Learning Content Object with the specified ID does not exist
  /api/v1/lco-model:
    get:
      tags:
      - Expert Module
      summary: Get Learning Content Object Models
      description: Get all Learning Content Object Models.
      operationId: getLearningContentObjectModels
      parameters:
      - $ref: '#/components/parameters/pageNumber'
      - $ref: '#/components/parameters/numberOfObjectsPerPage'
      responses:
        200:
          description: Successfully retrieved Learning Content Object Models
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LearningContentObjectModelList'
        400:
          description: Bad request
        401:
          description: Authentication failed
    post:
      tags:
      - Expert Module
      summary: Create Learning Content Object Model
      description: Create a metamodel that describes a Learning Content Object type.
      operationId: createLearningContentObjectModel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LearningContentObjectModel'
      responses:
        201:
          description: Successfully created Learning Content Object Model
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LearningContentObjectModel'
        400:
          description: Bad request
        401:
          description: Authentication failed
        409:
          description: A Learning Content Object Model of the specified type does already exist
        422:
          description: Unprocessable entity
      x-codegen-request-body-name: body
  /api/v1/lco-model/{lcoType}:
    get:
      tags:
      - Expert Module
      summary: Get Learning Content Object Model
      description: Get a metamodel that describes a Learning Content Object type.
      operationId: getLearningContentObjectModel
      parameters:
      - $ref: '#/components/parameters/lcoType'
      responses:
        200:
          description: Successfully retrieved Learning Content Object Model
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LearningContentObjectModel'
        401:
          description: Authentication failed
        404:
          description: A Learning Content Object Model for the specified type does not exist
    put:
      tags:
      - Expert Module
      summary: Update Learning Content Object Model
      description: Update a metamodel that describes a Learning Content Object type.
      operationId: updateLearningContentObjectModel
      parameters:
      - $ref: '#/components/parameters/lcoType'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LearningContentObjectModel'
      responses:
        200:
          description: Successfully updated Learning Content Object Model
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LearningContentObjectModel'
        400:
          description: Bad request
        401:
          description: Authentication failed
        404:
          description: A Learning Content Object Model for the specified type does not exist
        409:
          description: Learning Content Objects of the type defined by the metamodel do exist
        422:
          description: Unprocessable entity
      x-codegen-request-body-name: body
    delete:
      tags:
      - Expert Module
      summary: Delete Learning Content Object Model
      description: Delete a metamodel that describes a Learning Content Object type.
      operationId: deleteLearningContentObjectModel
      parameters:
      - $ref: '#/components/parameters/lcoType'
      responses:
        200:
          description: Successfully deleted Learning Content Object Model
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LearningContentObjectModel'
        401:
          description: Authentication failed
        404:
          description: A Learning Content Object Model for the specified type does not exist
        409:
          description: Learning Content Objects of the type defined by the metamodel do exist
...
components:
  schemas:
...
    PreliminaryLearningContentObject:
      type: object
      description: Schema for representing a generic Learning Content Object (LCO) to be created.
      required:
      - lcoType
      properties:
        lcoType:
          type: string
          description: The unique identifier of the type of the LCO.
          example: MODULE
        objectId:
          type: string
          description: The ID which is used to identify the LCO in an xAPI-statement.
          example: http://localhost:8081/goto.php?target=pg__86&client_id=default&h5p_object_id=1&obj_id_lrs=320
        attributes:
          type: array
          description: A list of attributes of the LCO.
          items:
            $ref: '#/components/schemas/LearningContentObjectAttribute'
    LearningContentObject:
      type: object
      description: Schema for representing a generic Learning Content Object (LCO).
      allOf:
      - $ref: '#/components/schemas/PreliminaryLearningContentObject'
      required:
      - lcoType
      - lcoId
      properties:
        lcoId:
          type: string
          description: The unique ID of this LCO.
          example: 633A587A-B944-4974-850F-DDDE69CECFF1
    LearningContentObjectList:
      type: object
      description: Schema for representing a container object that encapsulates Learning Content Objects.
      required:
      - lcos
      - totalNumber
      - providedNumber
      - pageNumber
      properties:
        lcos:
          type: array
          description: The encapsulated Learning Content Objects.
          items:
            $ref: '#/components/schemas/LearningContentObject'
        totalNumber:
          type: integer
          description: The total number of Learning Content Objects encapsulated.
          minimum: 0
        providedNumber:
          type: integer
          description: The provided number of Learning Content Objects.
          minimum: 0
        pageNumber:
          type: integer
          description: The provided page.
          minimum: 1
    LearningContentObjectModel:
      type: object
      description: Schema for representing a metamodel that describes a type of Learning Content Object.
      required:
      - lcoType
      properties:
        lcoType:
          type: string
          description: The unique identifier of the type of the LCO.
          example: MODULE
        attributes:
          type: array
          description: A list of attribute definitions of the LCO.
          items:
            $ref: '#/components/schemas/LearningContentObjectAttributeModel'
...
  parameters:
...
    lcoId:
      name: lcoId
      in: path
      description: Unique ID of a Learning Content Object.
      required: true
      schema:
        type: string
        example: 633A587A-B944-4974-850F-DDDE69CECFF1
...
    numberOfObjectsPerPage:
      name: objectsPerPage
      in: query
      description: The number of objects that should be delivered per page. If this is not specified, all objects are delivered. This has to be specified when a page number is specified.
      required: false
      schema:
        type: integer
        minimum: 1
    pageNumber:
      name: page
      in: query
      description: The page that should be delivered. The default value is one.
      required: false
      schema:
        type: integer
        minimum: 1
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT